+++ index.js
@@ -31,7 +31,9 @@ const _ = require('lodash');
 const { include } = require('@lykmapipo/include');
 const app = require('@lykmapipo/express-common');
 const pkg = include(__dirname, 'package.json');
+const AlertSource = include(__dirname, 'lib', 'source.model');
 const Alert = include(__dirname, 'lib', 'alert.model');
+const alertSourceRouter = include(__dirname, 'lib', 'source.http.router');
 const alertRouter = include(__dirname, 'lib', 'alert.http.router');
 
 
@@ -51,6 +53,18 @@ exports.info = _.merge({}, _.pick(pkg, [
 
 
 /**
+ * @name AlertSource
+ * @description AlertSource model
+ * @type {mongoose.Model}
+ *
+ * @author lally elias <lallyelias87@gmail.com>
+ * @since 1.0.0
+ * @version 0.1.0
+ */
+exports.AlertSource = AlertSource;
+
+
+/**
  * @name Alert
  * @description Alert model
  * @type {mongoose.Model}
@@ -60,6 +74,18 @@ exports.info = _.merge({}, _.pick(pkg, [
  * @version 0.1.0
  */
 exports.Alert = Alert;
+
+
+/**
+ * @name alertSourceRouter
+ * @description alert http router
+ * @type {express.Router}
+ *
+ * @author lally elias <lallyelias87@gmail.com>
+ * @since 1.0.0
+ * @version 0.1.0
+ */
+exports.alertSourceRouter = alertSourceRouter;
 
 
 /**
@@ -90,6 +116,7 @@ exports.apiVersion = alertRouter.apiVersion;
 Object.defineProperty(exports, 'app', {
   get() {
     /* @todo bind oauth middlewares authenticate, token, authorize */
+    app.mount(alertSourceRouter);
     app.mount(alertRouter);
     return app;
   }